// components/brave_rewards/resources/donate/constants/donate_types.ts
/* This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this file,
 * You can obtain one at http://mozilla.org/MPL/2.0/. */

export const enum types {
  ON_CLOSE_DIALOG = '@@rewards/ON_CLOSE_DIALOG',
  ON_PUBLISHER_BANNER = '@@rewards/ON_PUBLISHER_BANNER',
  ON_WALLET_PROPERTIES = '@@rewards/ON_WALLET_PROPERTIES',
  GET_WALLET_PROPERTIES = '@@rewards/GET_WALLET_PROPERTIES',
  ON_DONATE = '@@rewards/ON_DONATE',
  GET_RECURRING_TIPS = '@@rewards/GET_RECURRING_TIPS',
  ON_RECURRING_TIPS = '@@rewards/ON_RECURRING_TIPS',
  GET_RECONCILE_STAMP = '@@rewards/GET_RECONCILE_STAMP',
  ON_RECONCILE_STAMP = '@@rewards/ON_RECONCILE_STAMP',
  ON_RECURRING_TIP_REMOVED = '@@rewards/ON_RECURRING_TIP_REMOVED'
}